Complete the visible-canvas centering settings path
The centering UI needs a persisted AgentSettings flag and regression coverage so the setting can be saved, restored, and toggled without relying on unrelated local branch state.

Constraint: Keep this branch scoped to visible-canvas centering only
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep the visible-canvas centering flag inside AgentSettings and Canvas settings; do not mix it with experimental terminal flags
Tested: pnpm check; pnpm test -- --run tests/unit/contexts/agentSettings.spec.ts tests/unit/contexts/settingsPanel.spec.tsx
Not-tested: pnpm pre-commit E2E in the C: worktree (terminal runtime baseline is unstable there)
