+++ src/contexts/settings/domain/agentSettings.defaults.ts
@@ -1,0 +1,63 @@
+import { AGENT_PROVIDERS } from './agentSettings.providers'
+import type { AgentSettings } from './agentSettings'
+import { DEFAULT_UI_LANGUAGE } from './uiSettings'
+import { DEFAULT_WEBSITE_WINDOW_POLICY } from './websiteWindowSettings'
+
+export const DEFAULT_AGENT_SETTINGS: AgentSettings = {
+  language: DEFAULT_UI_LANGUAGE,
+  uiTheme: 'dark',
+  isPrimarySidebarCollapsed: false,
+  workspaceSearchPanelWidth: 420,
+  defaultProvider: 'codex',
+  agentProviderOrder: [...AGENT_PROVIDERS],
+  agentFullAccess: true,
+  defaultTerminalProfileId: null,
+  customModelEnabledByProvider: {
+    'claude-code': false,
+    codex: false,
+    opencode: false,
+    gemini: false,
+  },
+  customModelByProvider: {
+    'claude-code': '',
+    codex: '',
+    opencode: '',
+    gemini: '',
+  },
+  customModelOptionsByProvider: {
+    'claude-code': [],
+    codex: [],
+    opencode: [],
+    gemini: [],
+  },
+  taskTitleProvider: 'default',
+  taskTitleModel: '',
+  taskTagOptions: ['feature', 'bug', 'refactor', 'docs', 'test'],
+  taskPromptTemplates: [],
+  taskPromptTemplatesByWorkspaceId: {},
+  focusNodeOnClick: true,
+  focusNodeTargetZoom: 1,
+  focusNodeUseVisibleCanvasCenter: true,
+  standbyBannerEnabled: true,
+  standbyBannerShowTask: true,
+  standbyBannerShowSpace: true,
+  standbyBannerShowBranch: true,
+  standbyBannerShowPullRequest: true,
+  disableAppShortcutsWhenTerminalFocused: true,
+  keybindings: {},
+  canvasInputMode: 'auto',
+  canvasWheelBehavior: 'zoom',
+  canvasWheelZoomModifier: 'primary',
+  standardWindowSizeBucket: 'regular',
+  websiteWindowPolicy: DEFAULT_WEBSITE_WINDOW_POLICY,
+  experimentalWebsiteWindowPasteEnabled: false,
+  defaultTerminalWindowScalePercent: 80,
+  terminalFontSize: 13,
+  terminalFontFamily: null,
+  uiFontSize: 18,
+  githubPullRequestsEnabled: true,
+  updatePolicy: 'prompt',
+  updateChannel: 'stable',
+  releaseNotesSeenVersion: null,
+  hideWorktreeMismatchDropWarning: false,
+}
