+++ src/contexts/settings/domain/agentSettings.ts
@@ -19,13 +19,7 @@ import {
   type WorktreeNameSuggestionAgentProvider,
 } from './agentSettings.providers'
 import { normalizeFocusNodeTargetZoom, type FocusNodeTargetZoom } from './focusNodeTargetZoom'
-import {
-  DEFAULT_UI_LANGUAGE,
-  isValidUiLanguage,
-  isValidUiTheme,
-  type UiLanguage,
-  type UiTheme,
-} from './uiSettings'
+import { isValidUiLanguage, isValidUiTheme, type UiLanguage, type UiTheme } from './uiSettings'
 import {
   isValidUpdateChannel,
   isValidUpdatePolicy,
@@ -56,10 +50,8 @@ import {
   normalizeTaskPromptTemplates,
   normalizeTaskPromptTemplatesByWorkspaceId,
 } from './taskPromptTemplates'
-import {
-  DEFAULT_WEBSITE_WINDOW_POLICY,
-  normalizeWebsiteWindowPolicy,
-} from './websiteWindowSettings'
+import { normalizeWebsiteWindowPolicy } from './websiteWindowSettings'
+import { DEFAULT_AGENT_SETTINGS } from './agentSettings.defaults'
 
 export {
   FOCUS_NODE_TARGET_ZOOM_STEP,
@@ -139,6 +131,7 @@ export interface AgentSettings {
   taskPromptTemplatesByWorkspaceId: TaskPromptTemplatesByWorkspaceId
   focusNodeOnClick: boolean
   focusNodeTargetZoom: FocusNodeTargetZoom
+  focusNodeUseVisibleCanvasCenter: boolean
   standbyBannerEnabled: boolean
   standbyBannerShowTask: boolean
   standbyBannerShowSpace: boolean
@@ -162,64 +155,7 @@ export interface AgentSettings {
   releaseNotesSeenVersion: string | null
   hideWorktreeMismatchDropWarning: boolean
 }
-
-export const DEFAULT_AGENT_SETTINGS: AgentSettings = {
-  language: DEFAULT_UI_LANGUAGE,
-  uiTheme: 'dark',
-  isPrimarySidebarCollapsed: false,
-  workspaceSearchPanelWidth: 420,
-  defaultProvider: 'codex',
-  agentProviderOrder: [...AGENT_PROVIDERS],
-  agentFullAccess: true,
-  defaultTerminalProfileId: null,
-  customModelEnabledByProvider: {
-    'claude-code': false,
-    codex: false,
-    opencode: false,
-    gemini: false,
-  },
-  customModelByProvider: {
-    'claude-code': '',
-    codex: '',
-    opencode: '',
-    gemini: '',
-  },
-  customModelOptionsByProvider: {
-    'claude-code': [],
-    codex: [],
-    opencode: [],
-    gemini: [],
-  },
-  taskTitleProvider: 'default',
-  taskTitleModel: '',
-  taskTagOptions: ['feature', 'bug', 'refactor', 'docs', 'test'],
-  taskPromptTemplates: [],
-  taskPromptTemplatesByWorkspaceId: {},
-  focusNodeOnClick: true,
-  focusNodeTargetZoom: 1,
-  standbyBannerEnabled: true,
-  standbyBannerShowTask: true,
-  standbyBannerShowSpace: true,
-  standbyBannerShowBranch: true,
-  standbyBannerShowPullRequest: true,
-  disableAppShortcutsWhenTerminalFocused: true,
-  keybindings: {},
-  canvasInputMode: 'auto',
-  canvasWheelBehavior: 'zoom',
-  canvasWheelZoomModifier: 'primary',
-  standardWindowSizeBucket: 'regular',
-  websiteWindowPolicy: DEFAULT_WEBSITE_WINDOW_POLICY,
-  experimentalWebsiteWindowPasteEnabled: false,
-  defaultTerminalWindowScalePercent: 80,
-  terminalFontSize: 13,
-  terminalFontFamily: null,
-  uiFontSize: 18,
-  githubPullRequestsEnabled: true,
-  updatePolicy: 'prompt',
-  updateChannel: 'stable',
-  releaseNotesSeenVersion: null,
-  hideWorktreeMismatchDropWarning: false,
-}
+export { DEFAULT_AGENT_SETTINGS }
 
 function isValidTaskTitleProvider(value: unknown): value is TaskTitleProvider {
   return value === 'default' || isTaskTitleAgentProvider(value)
@@ -363,6 +299,9 @@ export function normalizeAgentSettings(value: unknown): AgentSettings {
     value.focusNodeTargetZoom,
     DEFAULT_AGENT_SETTINGS.focusNodeTargetZoom,
   )
+  const focusNodeUseVisibleCanvasCenter =
+    normalizeBoolean(value.focusNodeUseVisibleCanvasCenter) ??
+    DEFAULT_AGENT_SETTINGS.focusNodeUseVisibleCanvasCenter
   const standbyBannerEnabled =
     normalizeBoolean(value.standbyBannerEnabled) ?? DEFAULT_AGENT_SETTINGS.standbyBannerEnabled
   const standbyBannerShowTask =
@@ -468,6 +407,7 @@ export function normalizeAgentSettings(value: unknown): AgentSettings {
     taskPromptTemplatesByWorkspaceId,
     focusNodeOnClick,
     focusNodeTargetZoom,
+    focusNodeUseVisibleCanvasCenter,
     standbyBannerEnabled,
     standbyBannerShowTask,
     standbyBannerShowSpace,
