+++ tests/unit/contexts/agentSettings.spec.ts
@@ -34,4 +34,14 @@ describe('normalizeAgentSettings', () => {
       normalizeAgentSettings({ standardWindowSizeBucket: 'invalid' }).standardWindowSizeBucket,
     ).toBe(DEFAULT_AGENT_SETTINGS.standardWindowSizeBucket)
   })
+
+  it('defaults and normalizes the visible-canvas focus centering toggle', () => {
+    expect(DEFAULT_AGENT_SETTINGS.focusNodeUseVisibleCanvasCenter).toBe(true)
+    expect(normalizeAgentSettings({}).focusNodeUseVisibleCanvasCenter).toBe(true)
+    expect(
+      normalizeAgentSettings({
+        focusNodeUseVisibleCanvasCenter: false,
+      }).focusNodeUseVisibleCanvasCenter,
+    ).toBe(false)
+  })
 })
